+++ Changelog.md
@@ -1,6 +1,7 @@
 ## [Unreleased]
 
 - Fix: walk_dir_find_all_dir alogrithm update
+- Fix: _create_entry, _find_entry - walk through sectors
 
 ## [0.2.0] - 2024-08-29
 ### Fixed