+++ Changelog.md
@@ -4,6 +4,7 @@
 - Fix: _create_entry, _find_entry - walk through sectors
 - Future: add support permissions bit in classes
 - Add: mkdir, cmd_file_create_write add argument 'mode'
+- Future: add config permissoin mode for script
 
 
 ## [0.2.0] - 2024-08-29