+++ Changelog.md
@@ -1,5 +1,7 @@
 ## [Unreleased]
 
+## [0.3.0] - 2024-09-06
+
 - Fix: walk_dir_find_all_dir alogrithm update
 - Fix: _create_entry, _find_entry - walk through sectors
 - Future: add support permissions bit in classes